Add inlinable
